deployOnSonic now handles Timelock governance Split Sonic deploy up into vault/tokens and sonic staking strategy
(cherry picked from commit d57fd0c97e08c452f12d1e7f0a950348bb0e113c)
